Update DoubleClickPrevent + :3
