+++ Plugins/ColonThreeModule/plugin.json
@@ -1,0 +1,6 @@
+{
+    "name": "ColonThree",
+    "author": "1unar_Eclipse",
+    "version": ".3",
+    "description": ":3"
+}
