+++ Plugins/DoubleClickPrevent/plugin.json
@@ -1,6 +1,6 @@
 {
     "name": "DoubleClickPrevent",
     "author": "1unar_Eclipse",
-    "version": "0.2.0",
+    "version": "0.3.0",
     "description": "Adds artificial debounce time to prevent double (and further) clicks"
 }
